+++ contracts/admin/src/lib.rs
@@ -2,3 +2,5 @@ pub mod contract;
 pub mod execute;
 pub mod query;
 pub mod shared;
+#[cfg(test)]
+mod test;
